+++ Gems/EMotionFX/Code/EMotionFX/Source/SoftSkinDeformer.h
@@ -114,7 +114,7 @@ namespace EMotionFX
          * This does not alter the value returned by GetNumLocalBones().
          * @param numBones The number of bones to pre-allocate space for.
          */
-        MCORE_INLINE void ReserveLocalBones(uint32 numBones)                { mNodeNumbers.reserve(numBones); mBoneMatrices.reserve(numBones); }
+        MCORE_INLINE void ReserveLocalBones(size_t numBones)                { mNodeNumbers.reserve(numBones); mBoneMatrices.reserve(numBones); }
 
 
     protected:
